[AIRFLOW-3463] Move Log out of models.py (#4639)
