+++ airflow/utils/cli.py
@@ -30,7 +30,7 @@ import sys
 from argparse import Namespace
 from datetime import datetime
 
-import airflow.models
+from airflow.models.log import Log
 from airflow.utils import cli_action_loggers
 
 
@@ -46,7 +46,7 @@ def action_logging(f):
         end_datetime : end datetime instance by utc
         full_command : full command line arguments
         user : current user
-        log : airflow.models.Log ORM instance
+        log : airflow.models.log.Log ORM instance
         dag_id : dag id (optional)
         task_id : task_id (optional)
         execution_date : execution date (optional)
@@ -105,7 +105,7 @@ def _build_metrics(func_name, namespace):
     metrics['host_name'] = socket.gethostname()
 
     extra = json.dumps(dict((k, metrics[k]) for k in ('host_name', 'full_command')))
-    log = airflow.models.Log(
+    log = Log(
         event='cli_{}'.format(func_name),
         task_instance=None,
         owner=metrics['user'],
