+++ tests/api/common/experimental/test_delete_dag.py
@@ -30,7 +30,7 @@ from airflow.utils.state import State
 DM = models.DagModel
 DR = models.DagRun
 TI = models.TaskInstance
-LOG = models.Log
+LOG = models.log.Log
 TF = models.TaskFail
 TR = models.TaskReschedule
